initial commit MainActivity is the launcher activity. default fragment for this activity is ListFragment which show list of city. in the portrait orientation it just show one ListFragment but in the horizontal orientation its display two Fragment at the same time. one of them is ListFragment and another one is MapFragment. MapFragment is responsible for display location of the selected city on tha map and after click on the marker its open a popup who contain information about the city, users are able to use navigation in this step.
ListFragment : CitiesPresenterImpl is presenter for ListFragment and cityModelImpl class is responsible for fetch data list from json file assets and convert it to JsonArray with Gson library and notify presenter when data is ready.

cityModelImpl is also sort and filter data with Algorithm class.
In this class implemented merge algorithm for sort and binary search for filter.
final part for ListFragment is CitiesAdapter which it's responsible for show list item.

MapFragment : MapPresenterImpl is the presenter for this fragment.

Test Class :
android test:
CityPresenterImplTest for test CitiesPresenterImpl.
MainActivityInstrumentedTest for test MainActivity and orientations.
MapPresenterImplTest for test MapFragment.

java test:
AlgorithmUnitTest for test algorithm sort and search.

:x